Dlup is almost completely strictly typed
