+++ dlup/types.py
@@ -14,3 +14,4 @@ GenericNumber = Union[int, float]
 GenericNumberArray = Union[npt.NDArray[np.int_ | np.float_], Iterable[GenericNumber]]
 GenericFloatArray = Union[npt.NDArray[np.float_], Iterable[float]]
 GenericIntArray = Union[npt.NDArray[np.int_], Iterable[int]]
+ROIType = tuple[tuple[int, int], tuple[int, int]]
