+++ tests/test_dataset.py
@@ -6,14 +6,14 @@
 import numpy as np
 
 import dlup
-from dlup.data.dataset import TiledROIsSlideImageDataset, TilingMode
+from dlup.data.dataset import TiledWsiDataset, TilingMode
 
 
 def test_tiled_level_slide_image_dataset(monkeypatch, dlup_wsi):
     """Test a single image dataset."""
-    monkeypatch.setattr(TiledROIsSlideImageDataset, "slide_image", dlup_wsi)
+    monkeypatch.setattr(TiledWsiDataset, "slide_image", dlup_wsi)
     monkeypatch.setattr(dlup.SlideImage, "from_file_path", lambda x, backend: dlup_wsi)
-    dataset = TiledROIsSlideImageDataset.from_standard_tiling(
+    dataset = TiledWsiDataset.from_standard_tiling(
         "dummy",
         mpp=1.0,
         tile_size=(32, 24),
